requirements.txt
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
black==19.10b0
boto3==1.12.31
botocore==1.15.31
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
docutils==0.15.2
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jmespath==0.9.5
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.12.0
regex==2020.2.20
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
